Autenticação

+++ src/routes/memories.ts
@@ -3,8 +3,14 @@ import { z } from "zod";
 import { prisma } from "../lib/prisma";
 
 export async function memoriesroutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify();
+  });
+  app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: "asc",
       },
@@ -18,7 +24,7 @@ export async function memoriesroutes(app: FastifyInstance) {
     });
   });
 
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, reply) => {
     //const { id } = request.params;
     const paramsSchema = z.object({
       id: z.string().uuid(),
@@ -30,6 +36,10 @@ export async function memoriesroutes(app: FastifyInstance) {
         id,
       },
     });
+
+    if (!memory.isPublic && memory.userId != request.user.sub) {
+      return reply.status(401).send();
+    }
 
     return memory;
   });
@@ -46,13 +56,13 @@ export async function memoriesroutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "9ceba53f-4eee-450b-b27b-22a32bde04db",
+        userId: request.user.sub,
       },
     });
     return memory;
   });
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -67,7 +77,17 @@ export async function memoriesroutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (!memory.isPublic && memory.userId != request.user.sub) {
+      return reply.status(401).send();
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -80,12 +100,23 @@ export async function memoriesroutes(app: FastifyInstance) {
     return memory;
   });
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
 
     const { id } = paramsSchema.parse(request.params);
+
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (!memory.isPublic && memory.userId != request.user.sub) {
+      return reply.status(401).send();
+    }
+
     await prisma.memory.delete({
       where: {
         id,
